+++ components/NewsletterSubscribe/index.js
@@ -1,8 +1,8 @@
-import Mailchimp from '../Mailchimp';
+import ReCapchaWrapper from '../Recapcha';
 import NewsletterForm from '@/components/NewsletterSubscribe/NewsletterForm';
 
 const NewsletterSubscribe = () => {
-  return <Mailchimp child={<NewsletterForm />} />;
+  return <ReCapchaWrapper children={<NewsletterForm />} />;
 };
 
 export default NewsletterSubscribe;